+++ src/OrderManagement.Domain/Aggregates/Products/Events/ProductCreatedDomainEvent.cs
@@ -2,4 +2,4 @@ using OrderManagement.Domain.Core;
 
 namespace OrderManagement.Domain.Aggregates.Products;
 
-public record ProductCreatedDomainEvent(int Id) : IDomainEvent;
+public record ProductCreatedDomainEvent(long Id) : IDomainEvent;
